requirements.txt
flask==3.0.0
flask-cors==5.0.0
flask-jwt-extended==4.5.3
flask-pymongo==2.3.0
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.23.0 